Fixes issue with error display
Fixes issue with error display and if errors were set to display.
Changes display of errors to blank when errors are set to hidden. This
is different than the current message with shows a message to notify the
side administrator. This seems more consistent to the wording in the
dashboard the reads "Hide errors from site visitors"